+++ source/js/search-parameters/search-parameters.js
@@ -108,6 +108,14 @@ export class SearchParameters {
   }
 
   /**
+   * Returns root HTMLElement of the SearchParameters component
+   * @return {HTMLElement}
+   */
+  getHtmlElement() {
+    return document.getElementById(this.internalId);
+  }
+
+  /**
    * Removes all controls and related data
    */
   dispose() {
@@ -293,7 +301,7 @@ export class SearchParameters {
 
         if (!this.isController(this.searchParams[paramResourceType])) {
           getAutocompleterById(key).setList(
-            [paramName].concat(this.availableParams[paramResourceType])
+            [paramName].concat(this.availableParams[paramResourceType]).sort()
           );
         }
         if (key !== skipSearchItemId) {
@@ -317,9 +325,9 @@ export class SearchParameters {
           this.availableParams[resourceType].length) &&
         resourceType !== firstResourceType
     );
-    return (firstResourceType ? [firstResourceType] : []).concat(
-      resourceTypes.sort()
-    );
+    return (firstResourceType ? [firstResourceType] : [])
+      .concat(resourceTypes)
+      .sort();
   }
 
   /**
@@ -337,7 +345,7 @@ export class SearchParameters {
     <div class="section__body"></div>
 </div>
 <div>
-  <button id="${this.buttonId}" class="add-search-param-button">Add a search criterion</button>
+  <button id="${this.buttonId}" type="button" class="add-search-param-button">Add a search criterion</button>
 </div>`
     );
     document.getElementById(this.buttonId).onclick = () => this.addParam();
@@ -389,7 +397,7 @@ export class SearchParameters {
     <input type="text" id="${searchItemId}" value="${paramName}">
   </div>
   <div id="${searchItemContentId}" class="search-parameter__content"></div>
-  <button id="${removeButtonId}">remove</button>
+  <button id="${removeButtonId}" type="button">remove</button>
 </div>`
       );
     new Def.Autocompleter.Prefetch(paramResourceTypeSelectorId, [], {