+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [1.13.0] - 2020-09-08
+### Added
+- Display of minimum and maximum values that exist in the database
+  for a date range criterion
+- npm task "show-webpack-treemap" to run webpack-bundle-analyzer
+### Changed
+- Bundle size reduced by removing moment locales
+
 ## [1.12.0] - 2020-08-24
 ### Changed
 - Revised Observation criteria for Patient selection